feat: attempting to create a 3D model in rapier, will change depending on CAD
future goal is to use the CAD/OBJ model to do a simulation in python. There are multiple libraries that do this and it would be more accurate for our project. It could also help simulate the motor movement to achieve the undulation we want. It will not be able to account for fluid related resistance and interruption.